+++ Sticky/Features/Share/CardSlide.swift
@@ -28,8 +28,7 @@ struct CardSlide: View {
         let items = [
             Card(id: 0, name: "Hey"),
             Card(id: 1, name: "Ho"),
-            Card(id: 2, name: "Lets"),
-            Card(id: 3, name: "Go")
+            Card(id: 2, name: "Lets")
         ]
 
         return Carousel(
@@ -39,7 +38,7 @@ struct CardSlide: View {
         ) {
             // items를 돌며 생성
             ForEach(items, id: \.self.id) { item in
-                Item(
+                CardItem(
                     // 카드 UI의 id
                     id: Int(item.id),
                     // 카드 사이의 너비
@@ -62,129 +61,9 @@ struct CardSlide: View {
     }
 }
 
-// 카드 내에 정보를 담는 모델
-struct Card: Decodable, Hashable, Identifiable {
-    var id: Int
-    var name: String = ""
-}
-
 public class UIStateModel: ObservableObject {
     @Published var activeCard: Int = 0
     @Published var screenDrag: Float = 0.0
-}
-
-// 카드들이 담기는 Horizon 스크롤 뷰(?)같은 역할
-struct Carousel<Items: View>: View {
-    let items: Items
-    let numberOfItems: CGFloat // = 8
-    let spacing: CGFloat // = 16
-    let widthOfHiddenCards: CGFloat // = 32
-    let totalSpacing: CGFloat
-    let cardWidth: CGFloat
-
-    @GestureState var isDetectingLongPress = false
-
-    @EnvironmentObject var UIState: UIStateModel
-
-    @inlinable public init(
-        numberOfItems: CGFloat,
-        spacing: CGFloat,
-        widthOfHiddenCards: CGFloat,
-        @ViewBuilder _ items: () -> Items
-    ) {
-        self.items = items()
-        self.numberOfItems = numberOfItems
-        self.spacing = spacing
-        self.widthOfHiddenCards = widthOfHiddenCards
-        self.totalSpacing = (numberOfItems - 1) * spacing
-        self.cardWidth = UIScreen.main.bounds.width - (widthOfHiddenCards * 2) - (spacing * 2) // 279
-    }
-
-    var body: some View {
-        let totalCanvasWidth: CGFloat = (cardWidth * numberOfItems) + totalSpacing // 카드갯수 * 카드 너비 + (카드 갯수-1) * 카드 사이 너비
-        let xOffsetToShift = (totalCanvasWidth - UIScreen.main.bounds.width) / 2 // (총 너비 - 폰 너비) / 2
-        // 현재 보이는 카드가 화면 좌측 가장자리에서 떨어져있는 거리
-        let leftPadding = widthOfHiddenCards + spacing
-        let totalMovement = cardWidth + spacing
-
-        let activeOffset = xOffsetToShift + leftPadding - (totalMovement * CGFloat(UIState.activeCard))
-        let nextOffset = xOffsetToShift + leftPadding - (totalMovement * CGFloat(UIState.activeCard) + 1)
-
-        var calcOffset = Float(activeOffset)
-
-        if calcOffset != Float(nextOffset) {
-            calcOffset = Float(activeOffset) + UIState.screenDrag
-        }
-
-        return HStack(alignment: .center, spacing: spacing) {
-            items
-        }
-        .offset(x: CGFloat(calcOffset), y: 0)
-        .gesture(DragGesture().updating($isDetectingLongPress) { currentState, _, _ in
-            self.UIState.screenDrag = Float(currentState.translation.width)
-
-        }.onEnded { value in
-            self.UIState.screenDrag = 0
-
-            if value.translation.width < -50 {
-                if self.UIState.activeCard != Int(numberOfItems) - 1 {
-                    self.UIState.activeCard += 1
-                    let impactMed = UIImpactFeedbackGenerator(style: .medium)
-                    impactMed.impactOccurred()
-                }
-            }
-
-            if value.translation.width > 50 {
-                if self.UIState.activeCard != 0 {
-                    self.UIState.activeCard -= 1
-                    let impactMed = UIImpactFeedbackGenerator(style: .medium)
-                    impactMed.impactOccurred()
-                }
-            }
-        })
-    }
-}
-
-struct Canvas<Content: View>: View {
-    let content: Content
-    @EnvironmentObject var UIState: UIStateModel
-
-    @inlinable init(@ViewBuilder _ content: () -> Content) {
-        self.content = content()
-    }
-
-    var body: some View {
-        content
-            .frame(minWidth: 0, maxWidth: .infinity, minHeight: 0, maxHeight: .infinity, alignment: .center)
-            .background(Color.white.edgesIgnoringSafeArea(.all))
-    }
-}
-
-// 각 카드 정보가 담기는 UI 뷰
-struct Item<Content: View>: View {
-    @EnvironmentObject var UIState: UIStateModel
-    let cardWidth: CGFloat
-    let cardHeight: CGFloat
-    var id: Int
-    var content: Content
-
-    @inlinable public init(
-        id: Int,
-        spacing: CGFloat,
-        widthOfHiddenCards: CGFloat,
-        cardHeight: CGFloat,
-        @ViewBuilder _ content: () -> Content
-    ) {
-        self.content = content()
-        self.cardWidth = UIScreen.main.bounds.width - (widthOfHiddenCards * 2) - (spacing * 2)
-        self.cardHeight = cardHeight
-        self.id = id
-    }
-
-    var body: some View {
-        content
-            .frame(width: cardWidth, height: id == UIState.activeCard ? cardHeight : cardHeight - 60, alignment: .center)
-    }
 }
 
 struct CardSlide_Previews: PreviewProvider {
